员工信息与收入录入功能

+++ cis/src/main/java/cn/haohao/cis/user/service/IVuserIncomeService.java
@@ -44,7 +44,7 @@ public interface IVuserIncomeService {
 	 * @param vuserIncomeId
 	 * @return
 	 */
-	public VuserIncome getVuserIncomeById(Long vuserIncomeId);
+	public VuserIncome getVuserIncomeById(Integer vuserIncomeId);
 	/**
 	 * u-insert-b@note6.0@
 	 * u-insert-e@note6.0@
